collection/requirements.txt
black
grpcio
grpcio-tools
python-dotenv
pillow